+++ compiler/src/iree/compiler/Codegen/Common/Transforms.cpp
@@ -7,9 +7,16 @@
 #include "iree/compiler/Codegen/Common/Transforms.h"
 #include "iree/compiler/Codegen/Common/CombineLayoutTransformation.h"
 #include "iree/compiler/Dialect/LinalgExt/IR/LinalgExtOps.h"
+#include "iree/compiler/Dialect/LinalgExt/Transforms/Transforms.h"
+#include "iree/compiler/Dialect/Util/IR/UtilOps.h"
 #include "llvm/ADT/ScopeExit.h"
 #include "mlir/Analysis/SliceAnalysis.h"
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
+#include "mlir/Dialect/Arith/Utils/Utils.h"
+#include "mlir/Dialect/Linalg/IR/Linalg.h"
+#include "mlir/Dialect/MemRef/IR/MemRef.h"
+#include "mlir/Dialect/MemRef/Transforms/Transforms.h"
+#include "mlir/Dialect/Tensor/IR/Tensor.h"
 #include "mlir/Dialect/Utils/StaticValueUtils.h"
 
 #define DEBUG_TYPE "iree-codegen-common-transforms"
@@ -726,6 +733,101 @@ struct SwapCollapseShapeWithSlicePattern
 
 void populateSwapExtractWithCollapsePattern(RewritePatternSet &patterns) {
   patterns.add<SwapCollapseShapeWithSlicePattern>(patterns.getContext());
+}
+
+namespace {
+
+struct RemoveOptimizationBarrier final
+    : public OpRewritePattern<IREE::Util::OptimizationBarrierOp> {
+  using Base::Base;
+
+  LogicalResult matchAndRewrite(IREE::Util::OptimizationBarrierOp barrierOp,
+                                PatternRewriter &rewriter) const override {
+    rewriter.replaceOp(barrierOp, barrierOp.getOperands());
+    return success();
+  }
+};
+
+} // namespace
+
+std::optional<ReshapeOps> createDimensionExpansionOps(
+    RewriterBase &rewriter,
+    const llvm::SmallDenseMap<unsigned, int64_t> &expansionMap, Value v) {
+  auto tensorType = dyn_cast<RankedTensorType>(v.getType());
+  if (!tensorType) {
+    return std::nullopt;
+  }
+
+  SmallVector<OpFoldResult> outputShape;
+  SmallVector<ReassociationIndices> reassociation;
+  Location loc = v.getLoc();
+  SmallVector<OpFoldResult> origShape = tensor::getMixedSizes(rewriter, loc, v);
+
+  for (auto [index, dim] : llvm::enumerate(origShape)) {
+    reassociation.emplace_back(ReassociationIndices{});
+
+    if (!expansionMap.contains(index)) {
+      reassociation.back().push_back(outputShape.size());
+      outputShape.push_back(dim);
+      continue;
+    }
+
+    int64_t factor = expansionMap.lookup(index);
+    AffineExpr s0 = rewriter.getAffineSymbolExpr(0);
+    AffineExpr divExpr = s0.floorDiv(factor);
+    OpFoldResult newOuterDim = affine::makeComposedFoldedAffineApply(
+        rewriter, loc, divExpr, ArrayRef<OpFoldResult>{dim});
+    OpFoldResult newInnerDim = rewriter.getIndexAttr(factor);
+
+    reassociation.back().push_back(outputShape.size());
+    reassociation.back().push_back(outputShape.size() + 1);
+
+    outputShape.push_back(newOuterDim);
+    outputShape.push_back(newInnerDim);
+  }
+
+  auto staticOutputShape =
+      llvm::map_to_vector(outputShape, [](OpFoldResult ofr) {
+        if (auto staticShapeAttr = dyn_cast<Attribute>(ofr)) {
+          return cast<IntegerAttr>(staticShapeAttr).getInt();
+        }
+        return ShapedType::kDynamic;
+      });
+  auto outputType = RankedTensorType::get(
+      staticOutputShape, tensorType.getElementType(), tensorType.getEncoding());
+
+  auto expandShapeOp = tensor::ExpandShapeOp::create(
+      rewriter, loc, outputType, v, reassociation, outputShape);
+  Value barrier = IREE::Util::OptimizationBarrierOp::create(
+                      rewriter, loc, expandShapeOp.getResult())
+                      .getResult(0);
+  auto collapseShapeOp = tensor::CollapseShapeOp::create(
+      rewriter, loc, tensorType, barrier, reassociation);
+  return ReshapeOps{expandShapeOp, collapseShapeOp};
+}
+
+void populateRemoveOptimizationBarrierPatterns(RewritePatternSet &patterns) {
+  patterns.insert<RemoveOptimizationBarrier>(patterns.getContext());
+}
+
+void populateReshapePropagationPatterns(RewritePatternSet &patterns,
+                                        linalg::ControlFusionFn controlFn) {
+  if (!controlFn) {
+    controlFn = [](OpOperand *opOperand) {
+      return !isa_and_nonnull<linalg::FillOp, tensor::EmptyOp>(
+          opOperand->get().getDefiningOp());
+    };
+  }
+  linalg::populateFoldReshapeOpsByExpansionPatterns(patterns, controlFn);
+  IREE::LinalgExt::populateFoldReshapeOpsByExpansionPatterns(patterns,
+                                                             controlFn);
+  populateReshapeToInterfaceTensorPatterns(patterns);
+  populateCombineRelayoutOpPatterns(patterns);
+  populateFoldTensorReshapeIntoBufferPatterns(patterns);
+  tensor::populateFoldTensorEmptyPatterns(patterns);
+  tensor::populateBubbleUpExpandShapePatterns(patterns);
+  linalg::FillOp::getCanonicalizationPatterns(patterns, patterns.getContext());
+  memref::populateResolveRankedShapedTypeResultDimsPatterns(patterns);
 }
 
 } // namespace mlir::iree_compiler
